+++ apps/meteor/client/lib/omnichannel/businessHours/SingleBusinessHourBehavior.ts
