+++ apps/meteor/client/lib/omnichannel/livechatEnterprise/livechatSideNavItems.ts
@@ -1,5 +1,5 @@
-import { hasPermission, hasAtLeastOnePermission } from '../../../../client/lib/authorization';
-import { registerOmnichannelSidebarItem } from '../../../../client/views/omnichannel/sidebarItems';
+import { registerOmnichannelSidebarItem } from '../../../views/omnichannel/sidebarItems';
+import { hasPermission, hasAtLeastOnePermission } from '../../authorization';
 
 registerOmnichannelSidebarItem({
 	href: '/omnichannel/reports',
