+++ apps/meteor/client/views/omnichannel/crypto.ts
@@ -2,12 +2,6 @@ function ab2str(buf: ArrayLike<number> | ArrayBuffer): string {
 	return String.fromCharCode(...new Uint16Array(buf));
 }
 
-export async function generateKey(): Promise<string> {
-	const key = await crypto.subtle.generateKey({ name: 'AES-GCM', length: 256 }, true, ['encrypt', 'decrypt']);
-	const exportedKey = await crypto.subtle.exportKey('jwk', key);
-	return JSON.stringify(exportedKey);
-}
-
 export async function getKeyFromString(keyStr: string): Promise<CryptoKey> {
 	const key = JSON.parse(keyStr);
 	return crypto.subtle.importKey('jwk', key, { name: 'AES-GCM', length: 256 }, true, ['encrypt', 'decrypt']);
